+++ src/components/cc-addon-elasticsearch-options/cc-addon-elasticsearch-options.stories.js
@@ -84,4 +84,5 @@ export const preselectedKibana = makeStory(conf, {
 enhanceStoriesNames({
   defaultStory,
   noFlavorDetailsYet,
+  preselectedKibana,
 });