+++ src/components/cc-env-var-editor-json/cc-env-var-editor-json.js
@@ -143,6 +143,8 @@ export class CcEnvVarEditorJson extends LitElement {
           </div>
         ` : ''}
       <cc-input-text
+        label=${i18n('cc-env-var-editor-json.label')}
+        hidden-label
         multi
         clipboard
         value=${this._variablesAsJson}